+++ src/main/java/bucheon/leafy/application/event/request/LikeCancelEvent.java
@@ -1,22 +1,22 @@
-package bucheon.leafy.application.event;
+package bucheon.leafy.application.event.request;
 
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
-public class LikeEvent {
+public class LikeCancelEvent {
 
     private Long feedId;
     private Long userId;
 
     @Builder
-    private LikeEvent(Long feedId, Long userId) {
+    private LikeCancelEvent(Long feedId, Long userId) {
         this.feedId = feedId;
         this.userId = userId;
     }
 
-    public static LikeEvent of(Long feedId, Long userId) {
-        return LikeEvent.builder()
+    public static LikeCancelEvent of(Long feedId, Long userId) {
+        return LikeCancelEvent.builder()
                 .feedId(feedId)
                 .userId(userId)
                 .build();
