+++ src/main/java/bucheon/leafy/application/event/request/LikeEvent.java
@@ -1,22 +1,22 @@
-package bucheon.leafy.application.event;
+package bucheon.leafy.application.event.request;
 
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
-public class LikeCancelEvent {
+public class LikeEvent {
 
     private Long feedId;
     private Long userId;
 
     @Builder
-    private LikeCancelEvent(Long feedId, Long userId) {
+    private LikeEvent(Long feedId, Long userId) {
         this.feedId = feedId;
         this.userId = userId;
     }
 
-    public static LikeCancelEvent of(Long feedId, Long userId) {
-        return LikeCancelEvent.builder()
+    public static LikeEvent of(Long feedId, Long userId) {
+        return LikeEvent.builder()
                 .feedId(feedId)
                 .userId(userId)
                 .build();
